Add files via upload
Classification is probably the most popular task that you would deal with in real life. 
The below project is from Kaggle competition, We are going to create a classifier that predicts multiple features of the author of a given text. This is a Multilabel classification problem.

Dataset: This is a Kaggle competition project, details of data set as below, 
Blog Authorship Corpus
Over 600,000 posts from more than 19 thousand bloggers. The Blog Authorship Corpus consists of the collected posts of 19,320 bloggers gathered from blogger.com in August 2004. The corpus incorporates a total of 681,288 posts and over 140 million words - or approximately 35 posts and 7250 words per person. Each blog is presented as a separate file, the name of which indicates a blogger id# and the blogger’s self-provided gender, age, industry, and astrological sign. (All are labeled for gender and
age but for many, industry and/or sign is marked as unknown.)

Link to dataset: https://www.kaggle.com/rtatman/blog-authorship-corpus

Algorithm: 
1. Load the dataset, the dataset is large, use fewer rows.
2. Preprocess rows of the “text” column
a. Remove unwanted characters
b. Convert text to lowercase
c. Remove unwanted spaces
d. Remove stopwords
3. As we want to make this into a multi-label classification problem, you are required to merge all the label columns together, so that we have all the labels together for a particular sentence, Label columns to merge: “gender”, “age”, “topic”, “sign”
4. Separate features and labels, and split the data into training and testing
5. Vectorize the features - a Bag of Words using count vectorizer
6. Create a dictionary to get the count of every label i.e. the key will be label name and value will
be the total count of the label. Check below image for reference
7. Transform the labels - As we have noticed before, in this task each example can have multiple tags. To deal with such kind of prediction, we need to transform labels in a binary form and the prediction will be mask of 0s and 1s. For this purpose, it is convenient to use MultiLabelBinarizer from sklearn
a. Convert your train and test labels using MultiLabelBinarizer
8. Choose a classifier - using the One-vs-Rest approach, which is implemented in OneVsRestClassifier class. In this approach k classifiers (= number of tags) are trained. As a basic classifier, use LogisticRegression . It is one of the simplest methods, but often it performs good enough in text classification tasks. It might take some time because the number of classifiers to train is large.
9. Fit the classifier, make predictions and get the accuracy
10. Print true label and predicted label for any five examples